[JL] discovered test cleanups cannot work lol

+++ server/tests/unit-tests/question-api.test.ts
@@ -47,26 +47,29 @@ describe("My test", () => {
     })
   })
 
+  // Cleanup is not possible as the server shuts down automatically right before this step...
   afterAll(async () => {
     try {
-      await $fetch(`/api/users/${testUser.uid}`, {
-        method: "DELETE",
-      })
-    } catch (e) {}
-    try {
-      await $fetch(`/api/mycourses/${testCourse.id}`, {
-        method: "DELETE",
-      })
-    } catch (e) {}
-    try {
-      await $fetch(`/api/topics/${testTopic.id}`, {
-        method: "DELETE",
-      })
-    } catch (e) {}
-    try {
-      await $fetch(`/api/quiz/${testQuiz.id}`, {
-        method: "DELETE",
-      })
+      if (testUser) {
+        await $fetch(`/api/users/${testUser.uid}`, {
+          method: "DELETE",
+        })
+      }
+      if (testCourse) {
+        await $fetch(`/api/mycourses/${testCourse.id}`, {
+          method: "DELETE",
+        })
+      }
+      if (testTopic) {
+        await $fetch(`/api/topics/${testTopic.id}`, {
+          method: "DELETE",
+        })
+      }
+      if (testQuiz) {
+        await $fetch(`/api/quiz/${testQuiz.id}`, {
+          method: "DELETE",
+        })
+      }
     } catch (e) {}
   })
